# BarToYou-WebApi/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BarToYou API",
        "description": "Documentación Swagger de la API de BarToYou",
        "contact": {
            "email": "soporte@bartoyou.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/bartoyou/consumption-recipes": {
            "get": {
                "tags": [
                    "Consumption Recipes"
                ],
                "summary": "Listar todas las recetas de consumo",
                "operationId": "438b126d08b89c809822fac869702e46",
                "responses": {
                    "200": {
                        "description": "Lista de recetas paginada"
                    }
                }
            },
            "post": {
                "tags": [
                    "Consumption Recipes"
                ],
                "summary": "Crear una nueva receta de consumo",
                "operationId": "fefa50cfc223438ff673da2a3a1d242c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "consumption_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "ingredient_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "ingredient_amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 50
                                    },
                                    "ingredient_unit": {
                                        "type": "string",
                                        "example": "ml"
                                    },
                                    "custom_drink_id": {
                                        "type": "string",
                                        "example": "#1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Receta creada correctamente"
                    },
                    "400": {
                        "description": "La receta ya existe"
                    }
                }
            }
        },
        "/api/bartoyou/consumption-recipes/{id}": {
            "get": {
                "tags": [
                    "Consumption Recipes"
                ],
                "summary": "Obtener una receta de consumo por ID",
                "operationId": "a295bc7be6966ae67ab255d227526341",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la receta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Receta encontrada"
                    },
                    "404": {
                        "description": "Receta no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Consumption Recipes"
                ],
                "summary": "Actualizar una receta de consumo",
                "operationId": "044126e04ecfc88ad12de1abff28a076",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la receta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ingredient_amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 60
                                    },
                                    "ingredient_unit": {
                                        "type": "string",
                                        "example": "ml"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Receta actualizada correctamente"
                    },
                    "404": {
                        "description": "Receta no encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Consumption Recipes"
                ],
                "summary": "Eliminar una receta de consumo",
                "operationId": "46a463198801d239743711791a3a3728",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la receta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Receta eliminada correctamente"
                    },
                    "404": {
                        "description": "Receta no encontrada"
                    }
                }
            }
        },
        "/api/custom-drink": {
            "post": {
                "tags": [
                    "Custom Drink"
                ],
                "summary": "Crear una bebida personalizada",
                "operationId": "91e47356907f72661d1bffc4a34a8fa6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "base_drink",
                                    "base_drink_id",
                                    "ingredients"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "base_drink": {
                                        "type": "string",
                                        "example": "Ron"
                                    },
                                    "base_drink_id": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "ingredients": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "ingredient_id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "amount": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 30.5
                                                },
                                                "unit": {
                                                    "type": "string",
                                                    "example": "ml"
                                                },
                                                "consumption_id": {
                                                    "type": "integer",
                                                    "example": 5
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "ice": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "ice_type": {
                                        "type": "string",
                                        "example": "picado"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bebida personalizada creada correctamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
...
    },
...
}